handle when same componentize-py.toml appears twice in PYTHON_PATH
In this case, we need to decide which directory "owns" the file, which affects
how we derive package names relative to that directory.  For example, if both
`foo` and `foo/.venv/lib/python3.12/site-packages` are in `PYTHON_PATH`, and we
find a componentize-py.toml file under the latter, then we'll also find it under
the former.  In that case, we consider the latter to be the "owner" of the file.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>