append `Raises` line to function docstrings as appropriate
If the WIT representation of a function returns `result<A,B>`, then we generate
a Python function that returns `A` and raises `Err(B)`.  Since there's currently
no way of representing the `Err(B)` part in the function's signature, we do the
next best thing: put it in the docstring.

Fixes #53

Signed-off-by: Joel Dice <joel.dice@fermyon.com>