rename "lift" and "lower" functions
The "lift" and "lower" terminology was misleading and confusing for anyone
familiar with their meanings in the Component Model ABI.  I was overloading them
to mean "lift from the canonical ABI representation to Python" and "lower from
Python to the canonical ABI representation", respectively, but that's easy to
confuse with how the canonical ABI uses them to mean "lift a core function to a
component function" and "lower a component function to a core function".

In order to avoid such confusion, I've renamed them to "from-canon" and
"to-canon", respectively.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>