(re)add `--stub-wasi` option
Early versions of this project had such an option, but I had to drop as part of
the refactoring required to support native extensions.  The implementation is a
bit tricky, since we need to import WASI during pre-initialization, but then
stub out the imports in the final component.  Normally, that would require
significant surgery on the pre-initialized component, but I've "cheated" by
applying the state snapshot to a stubbed component instead of the original one.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>