fix unresolved symbol issue when using latest Rust `nightly`
As of this writing, recent Rust `nightly` builds include a version of the `libc`
crate that expects `wasi-libc` to define the following global variables, but
`wasi-libc` defines them as preprocessor constants which aren't visible at link
time, so we need to define them somewhere.  Ideally, we should fix this
upstream, but for now we work around it.

This commit also addresses an unused field warning.

Fixes #86

Signed-off-by: Joel Dice <joel.dice@fermyon.com>