fix resource test regressions
Py03 0.22.x no longer has `Py::into_ref`, which returned a reference that was
guaranteed to survive until the GIL was released.  We were relying on that to
ensure borrow handles were released before returning to the caller.  Since
there's no equivalent feature in 0.22.x AFAICT, we need to clean up borrow
handles ourselves.

This also fixes `componentize_py_to_canon_handle`, which was causing reference
counts to be decremented too early.

Note that, although the tests are now passing, there may still be leaks due to
reference counts not being decremented where they should.  Not sure what the
best way to test that might be; we might need to research Python tools for heap
profiling and leak detection.

Finally, I haven't tested the examples yet; we'll want to do that before we
merge any of this into main.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>