update and fix examples (#120)
An earlier commit which updated various dependencies (most notably, Wasmtime)
unmasked a bug in `add_wasi_and_stubs` such that we were overwriting the host
functions added by `wasmtime-wasi` with trapping stubs.  This wasn't a problem
previously because there was only one version of WASIp2: 0.2.0.  Now that we
have 0.2.1, 0.2.2, etc., with corresponding support in Wasmtime, we need to be a
bit more careful.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>