fix memory leaks in generated bindings (#153)
* fix memory leaks in generated bindings

This addresses a couple of issues:

- We weren't freeing canonical ABI results after calling imports and then
  converting the results to Python values.

- We were using the PyO3 `Bound` APIs incorrectly in a few places, resulting in
  refcounts being incremented too many times.

While debugging the latter issue, I took the opportunity to update to the latest
PyO3 release, which required addressing the usual API breakage.  It also
required temporarily forking the repo to add a patch to disable `extern
"C-unwind"` in `pyo3_ffi` on `wasm32`.  We should be able to drop that fork once
we upgrade everything to Python 3.14.

Finally, I discovered
https://pyo3.rs/v0.25.0/features.html#pyo3_disable_reference_pool which
supposedly optimizes Rust<->Python transitions by disabling a feature we don't
rely on anyway, so I went ahead and enabled it.

Fixes #152

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* bump version to 0.17.1

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* update python.rs to use new PyO3 API

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>