add async and WASIp3 support (#173)
This adds basic support for async exports, async imports, streams, and futures.
See the new `http-p3` example for a practical demonstration.

This does not include support for cancelling subtasks, reads, or writes.  That
shouldn't be difficult to add, but can wait until a later PR.

I've also used this opportunity to update the Rust edition to 2024, update some
deps, and reflow some comments.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>