fix dropping of resource borrows in async-lifted exports (#182)
* fix dropping of resource borrows in async-lifted exports

Previously, I had been using a global variable to track borrows which needed to
be dropped when returning from an export, which works fine for sync-lifted
exports but not so fine for async-lifted ones.  In the latter case, we would
incorrectly drop borrows from other concurrent calls.  Now I've eliminated the
global variable in favor of tracking the borrows on a per-call basis.

This also includes few minor chores:

- Switch CI to use final CPython 3.14.0 release
- Update Wasmtime to v39.0.0
- Update to newer wasm-tools commit

I had also intended to update to WASI-SDK 29, but it seems to have broken the
TCP example.  Will debug that later.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* specify wasmtime==38.0.0 when testing sandbox example

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>